Move project 11 files to samples folder
