+++ projects/src/samples/project_11/average.ts
