+++ projects/src/samples/project_11/seven.ts
