+++ projects/src/samples/project_11/square/main.ts
