+++ projects/src/samples/project_11/square/square.ts
