+++ simulator/src/languages/jack.test.ts
@@ -1,5 +1,5 @@
 import { unwrap } from "@davidsouther/jiffies/lib/esm/result";
-import { Programs } from "@nand2tetris/projects/project_11/index.js";
+import { Programs } from "@nand2tetris/projects/samples/project_11/index.js";
 import { JACK } from "./jack";
 
 describe("compiler", () => {
